+++ sdk/keyvault/keyvault-admin/samples/v4/typescript/src/accessControlHelloWorld.ts
@@ -67,11 +67,9 @@ export async function main(): Promise<void> {
   assignment = await client.getRoleAssignment(globalScope, roleAssignmentName);
   console.log(assignment);
 
-  assignment = await client.deleteRoleAssignment(globalScope, roleAssignmentName);
-  console.log(assignment);
+  await client.deleteRoleAssignment(globalScope, roleAssignmentName);
 
-  roleDefinition = await client.deleteRoleDefinition(globalScope, roleDefinition.name);
-  console.log(roleDefinition);
+  await client.deleteRoleDefinition(globalScope, roleDefinition.name);
 }
 
 main().catch((err) => {