+++ sdk/keyvault/keyvault-admin/src/accessControlClient.ts
@@ -166,8 +166,7 @@ export class KeyVaultAccessControlClient {
    * ```ts
    * const client = new KeyVaultAccessControlClient(url, credentials);
    * const roleAssignment = await client.createRoleAssignment("/", "295c179b-9ad3-4117-99cd-b1aa66cf4517");
-   * const deletedRoleAssignment = await client.deleteRoleAssignment(roleAssignment.properties.roleScope, roleAssignment.name);
-   * console.log(deletedRoleAssignment);
+   * await client.deleteRoleAssignment(roleAssignment.properties.roleScope, roleAssignment.name);
    * ```
    * Deletes an existing role assignment.
    * @param roleScope - The scope of the role assignment.
@@ -178,15 +177,9 @@ export class KeyVaultAccessControlClient {
     roleScope: KeyVaultRoleScope,
     name: string,
     options: DeleteRoleAssignmentOptions = {}
-  ): Promise<KeyVaultRoleAssignment> {
+  ): Promise<void> {
     return withTrace("deleteRoleAssignment", options, async (updatedOptions) => {
-      const response = await this.client.roleAssignments.delete(
-        this.vaultUrl,
-        roleScope,
-        name,
-        updatedOptions
-      );
-      return mappings.roleAssignment.generatedToPublic(response);
+      await this.client.roleAssignments.delete(this.vaultUrl, roleScope, name, updatedOptions);
     });
   }
 
@@ -482,9 +475,8 @@ export class KeyVaultAccessControlClient {
    * Example usage:
    * ```ts
    * const client = new KeyVaultAccessControlClient(url, credentials);
-   * const roleDefinition = await client.upsertRoleDefintion("/", "23b8bb1a-39c0-4c89-a85b-dd3c99273a8a", []);
-   * const deletedRoleDefinition = await client.deleteRoleDefinition("/", roleDefinition.name);
-   * console.log(deletedRoleDefinition);
+   * const roleDefinition = await client.setRoleDefinition("/", "23b8bb1a-39c0-4c89-a85b-dd3c99273a8a", []);
+   * await client.deleteRoleDefinition("/", roleDefinition.name);
    * ```
    * @param roleScope - The scope of the role definition.
    * @param name - The name of the role definition to delete.
@@ -494,15 +486,9 @@ export class KeyVaultAccessControlClient {
     roleScope: KeyVaultRoleScope,
     name: string,
     options: DeleteRoleDefinitionOptions = {}
-  ): Promise<KeyVaultRoleDefinition> {
+  ): Promise<void> {
     return withTrace("deleteRoleDefinition", options, async (updatedOptions) => {
-      const response = await this.client.roleDefinitions.delete(
-        this.vaultUrl,
-        roleScope,
-        name,
-        updatedOptions
-      );
-      return mappings.roleDefinition.generatedToPublic(response);
+      await this.client.roleDefinitions.delete(this.vaultUrl, roleScope, name, updatedOptions);
     });
   }
 }