+++ sdk/metricsadvisor/ai-metrics-advisor/samples/v1/typescript/src/dataSourceCredential.ts
@@ -110,7 +110,8 @@ async function updateDatasourceCredential(
 
   try {
     console.log(`Updating credential ${credentialId}...`);
-    const updated = await client.updateDatasourceCredential(credentialId, patch);
+    await client.updateDatasourceCredential(credentialId, patch);
+    const updated = await client.getDataFeed(credentialId);
     console.dir(updated);
   } catch (err) {
     console.log("Error occurred when updating credential");