+++ sdk/servicebus/service-bus/test/internal/unit/abortSignal.spec.ts
@@ -20,7 +20,6 @@ import {
   createConnectionContextForTestsWithSessionId
 } from "./unittestUtils";
 import { StandardAbortMessage } from "@azure/core-amqp";
-import { isLinkLocked } from "../utils/misc";
 import { ServiceBusSessionReceiverImpl } from "../../../src/receivers/sessionReceiver";
 import { ServiceBusReceiverImpl } from "../../../src/receivers/receiver";
 import { MessageSession } from "../../../src/session/messageSession";
@@ -156,6 +155,38 @@ describe("AbortSignal", () => {
         assert.isTrue((err as any).retryable);
       }
     });
+
+    it("_trySend passes abortSignal to awaitable sender", async () => {
+      const sender = new MessageSender(connectionContext, "fakeEntityPath", {
+        timeoutInMs: 1
+      });
+      closeables.push(sender);
+
+      let wasAbortSignalPassed = false;
+      sender["_link"] = {
+        credit: 999,
+        isOpen: () => true,
+        session: {
+          outgoing: {
+            available: () => true
+          }
+        },
+        sendable() {
+          return true;
+        },
+        send(_msg, options) {
+          if (options?.abortSignal) {
+            wasAbortSignalPassed = true;
+          }
+          return Promise.resolve({});
+        }
+      } as AwaitableSender;
+
+      await sender["_trySend"]({} as Buffer, true, {
+        abortSignal: createAbortSignalForTest(false)
+      });
+      assert.isTrue(wasAbortSignalPassed, "abortSignal should have been passed to AwaitableSender");
+    });
   });
 
   describe("MessageSender.open() aborts after...", () => {
@@ -172,8 +203,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(sender));
     });
 
     it("...afterLock", async () => {
@@ -189,8 +218,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(sender));
     });
 
     it("...negotiateClaim", async () => {
@@ -219,8 +246,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(sender));
     });
 
     it("...createAwaitableSender", async () => {
@@ -249,8 +274,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(sender));
     });
   });
 
@@ -272,8 +295,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(messageReceiver));
     });
 
     it("...after negotiateClaim", async () => {
@@ -298,8 +319,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(messageReceiver));
     });
 
     it("...after createReceiver", async () => {
@@ -325,8 +344,6 @@ describe("AbortSignal", () => {
         assert.equal(err.message, StandardAbortMessage);
         assert.equal(err.name, "AbortError");
       }
-
-      assert.isFalse(isLinkLocked(messageReceiver));
     });
   });