+++ sdk/tables/data-tables/recordings/node/tableclient_sasconnectionstring_createentity_getentity_and_delete/recording_should_createentity_with_guid.js
@@ -1,0 +1,87 @@
+let nock = require('nock');
+
+module.exports.hash = "a66a20d62f23f75ef5943d4f183e331b";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":true})
+  .post('/tableClientTestSASConnectionStringnode', {"PartitionKey":"P3_node","RowKey":"R3","testField":"cf8ef051-1b7d-4e93-a1e5-a3944d7e441c","testField@odata.type":"Edm.Guid"})
+  .query(true)
+  .reply(204, "", [
+  'Cache-Control',
+  'no-cache',
+  'Content-Length',
+  '0',
+  'ETag',
+  `W/"datetime'2021-06-09T16%3A21%3A45.2502545Z'"`,
+  'Location',
+  "https://fakeaccount.table.core.windows.net/tableClientTestSASConnectionStringnode(PartitionKey='P3_node',RowKey='R3')",
+  'Server',
+  'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '654ef09a-9002-001e-4f4b-5d5b52000000',
+  'x-ms-client-request-id',
+  '75ac6336-cfb7-4666-a86c-37e7bab89905',
+  'x-ms-version',
+  '2019-02-02',
+  'X-Content-Type-Options',
+  'nosniff',
+  'Preference-Applied',
+  'return-no-content',
+  'DataServiceId',
+  "https://fakeaccount.table.core.windows.net/tableClientTestSASConnectionStringnode(PartitionKey='P3_node',RowKey='R3')",
+  'Date',
+  'Wed, 09 Jun 2021 16:21:45 GMT'
+]);
+
+nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":true})
+  .get(`/tableClientTestSASConnectionStringnode(PartitionKey='P3_node',RowKey='R3')`)
+  .query(true)
+  .reply(200, {"odata.metadata":"https://fakeaccount.table.core.windows.net/$metadata#tableClientTestSASConnectionStringnode/@Element","odata.etag":"W/\"datetime'2021-06-09T16%3A21%3A45.2502545Z'\"","PartitionKey":"P3_node","RowKey":"R3","Timestamp":"2021-06-09T16:21:45.2502545Z","testField@odata.type":"Edm.Guid","testField":"cf8ef051-1b7d-4e93-a1e5-a3944d7e441c"}, [
+  'Cache-Control',
+  'no-cache',
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json;odata=minimalmetadata;streaming=true;charset=utf-8',
+  'ETag',
+  `W/"datetime'2021-06-09T16%3A21%3A45.2502545Z'"`,
+  'Server',
+  'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '654ef0af-9002-001e-5e4b-5d5b52000000',
+  'x-ms-client-request-id',
+  'd677dfc0-f03d-450c-b670-886b3d620444',
+  'x-ms-version',
+  '2019-02-02',
+  'X-Content-Type-Options',
+  'nosniff',
+  'Access-Control-Expose-Headers',
+  'x-ms-request-id,x-ms-client-request-id,Server,x-ms-version,X-Content-Type-Options,Cache-Control,ETag,Content-Type,Content-Length,Date,Transfer-Encoding',
+  'Access-Control-Allow-Origin',
+  '*',
+  'Date',
+  'Wed, 09 Jun 2021 16:21:45 GMT'
+]);
+
+nock('https://fakeaccount.table.core.windows.net:443', {"encodedQueryParams":true})
+  .delete(`/tableClientTestSASConnectionStringnode(PartitionKey='P3_node',RowKey='R3')`)
+  .query(true)
+  .reply(204, "", [
+  'Cache-Control',
+  'no-cache',
+  'Content-Length',
+  '0',
+  'Server',
+  'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
+  'x-ms-request-id',
+  '654ef0b9-9002-001e-664b-5d5b52000000',
+  'x-ms-client-request-id',
+  '7aa02795-b9ab-4396-9a8f-a11894a016f7',
+  'x-ms-version',
+  '2019-02-02',
+  'X-Content-Type-Options',
+  'nosniff',
+  'Date',
+  'Wed, 09 Jun 2021 16:21:45 GMT'
+]);